Mods to run UMWM in GEOS
The bulk of these changes are related to using 2D decomposition
of the global domain. UMWM is effectively run in regional
mode to model waves in each of the GEOS tiles. This approach
requires 1-point halo and thus a slight increase of the size
of the actual computational domain corresponding to one of
the tiles.

Additional work is required to ensure reproducibility.

Stokes drift velocity is currently not used in the calculation
of the relative wind speed.